# Image Editor: replace window.onresize and add transition animation