--- a/client/blocks/image-editor/image-editor-canvas.jsx
+++ b/client/blocks/image-editor/image-editor-canvas.jsx
@@ -4,7 +4,7 @@
 import React, { Component, PropTypes } from 'react';
 import ReactDom from 'react-dom';
 import { connect } from 'react-redux';
-import { noop, throttle, startsWith } from 'lodash';
+import { noop, startsWith } from 'lodash';
 import classNames from 'classnames';
 
 /**
@@ -62,14 +62,19 @@ class ImageEditorCanvas extends Component {
 		isImageLoaded: false
 	};
 
+	// throttle the frame rate of window.resize() to circa 30fps
+	frameRateInterval = 1000 / 30;
+	requestAnimationFrameId = null;
+	lastTimestamp = null;
+
+	onWindowResize = () => {
+		this.requestAnimationFrameId = window.requestAnimationFrame( this.updateCanvasPosition );
+	};
+
 	constructor( props ) {
 		super( props );
-
-		this.onWindowResize = null;
-
 		this.onLoadComplete = this.onLoadComplete.bind( this );
 		this.updateCanvasPosition = this.updateCanvasPosition.bind( this );
-
 		this.isVisible = false;
 	}
 
@@ -126,8 +131,9 @@ class ImageEditorCanvas extends Component {
 
 		this.drawImage();
 		this.updateCanvasPosition();
-		this.onWindowResize = throttle( this.updateCanvasPosition, 200 );
+
 		if ( typeof window !== 'undefined' ) {
+			this.lastTimestamp = window.performance.now();
 			window.addEventListener( 'resize', this.onWindowResize );
 		}
 
@@ -137,7 +143,7 @@ class ImageEditorCanvas extends Component {
 	componentWillUnmount() {
 		if ( typeof window !== 'undefined' && this.onWindowResize ) {
 			window.removeEventListener( 'resize', this.onWindowResize );
-			this.onWindowResize = null;
+			window.cancelAnimationFrame( this.requestAnimationFrameId );
 		}
 
 		this.isVisible = false;
@@ -221,7 +227,18 @@ class ImageEditorCanvas extends Component {
 		context.restore();
 	}
 
-	updateCanvasPosition() {
+	updateCanvasPosition( timestamp ) {
+		const now = timestamp,
+			elapsedTime = now - this.lastTimestamp;
+
+		if ( elapsedTime < this.frameRateInterval ) {
+			return;
+		}
+
+		// if enough time has passed to call the next frame
+		// reset lastTimeStamp minus 1 frame in ms ( to adjust for frame rates other than 60fps )
+		this.lastTimestamp = now - ( elapsedTime % this.frameRateInterval );
+
 		const {
 			leftRatio,
 			topRatio,
@@ -233,11 +250,13 @@ class ImageEditorCanvas extends Component {
 			canvasX = -50 * widthRatio - 100 * leftRatio,
 			canvasY = -50 * heightRatio - 100 * topRatio;
 
+		const { offsetTop, offsetLeft, offsetWidth, offsetHeight } = canvas;
+
 		this.props.setImageEditorCropBounds(
-			canvas.offsetTop - canvas.offsetHeight * -canvasY / 100,
-			canvas.offsetLeft - canvas.offsetWidth * -canvasX / 100,
-			canvas.offsetTop + canvas.offsetHeight * ( 1 + canvasY / 100 ),
-			canvas.offsetLeft + canvas.offsetWidth * ( 1 + canvasX / 100 )
+			offsetTop - offsetHeight * -canvasY / 100,
+			offsetLeft - offsetWidth * -canvasX / 100,
+			offsetTop + offsetHeight * ( 1 + canvasY / 100 ),
+			offsetLeft + offsetWidth * ( 1 + canvasX / 100 )
 		);
 	}