--- a/client/blocks/image-editor/style.scss
+++ b/client/blocks/image-editor/style.scss
@@ -55,7 +55,6 @@
 	position: absolute;
 	top: 50%;
 	left: 50%;
-
 	&.is-placeholder {
 		@include placeholder-dark();
 	}
@@ -65,6 +64,7 @@
 	position: absolute;
 	border: 1px solid $white;
 	cursor: move;
+	box-sizing: border-box;
 }
 
 .image-editor__crop-background {