@gwwar 2017-07-10
Overall I think this makes sense 👍 

I'm seeing similar performance on my laptop, but I'll try again soon on a slower machine. If you already had a profile of the original/new behavior it'd be handy to add as a comment. Maybe something similar to https://github.com/Automattic/wp-calypso/pull/10836#issuecomment-274636273 but with us dragging the crop controls when we begin recording.