@ramonjd 2017-07-11
Thanks for the feedback, everyone. :)

Final improvements according to comments, and squashed some commits.

I also spent some time recording performance. 

On all the browsers available to me, there was no huge difference in rendering time between window.resize.

However once I started CPU throttling, I began to see some benefits of raf/throttling on total rendering time.

Tests were done using a 1000 x 800 image and resizing the window in and out 3 x for 4secs each. 

**Without raf**
Browser spent 6.5secs rendering. 

**With raf throtting set to 30fps**
Browser spent 3secs rendering. 

**With raf throtting set to 45fps**
5.9 seconds

The only differences was that _.throttling the window.resize event appears to come with better garbage collection. Chrome only does a dump once you stop resizing.

That the cropping area's border doesn't have to play as much 'catch up' onResize with the canvas element, I submit, is a worthy compromise.