@ramonjd 2017-07-11
@marcinbot @retrofox Easy change to use destructuring assignment, if you think it's clearer to read. Thanks.
I'll take another look at reducing the bouncing effect as you mentioned. 👍