@ramonjd 2017-07-14
The 'jump' effect was caused by the fraction-of-a-second delay between the canvas transform and the background box resize/repositioning.

I couldn't find a simple, clean way to sync the transition, so reverting it, and labelling it 'not worth it right now'.